more misc cleanup
fixed some bool checks and simplified a few if/elif chains. I'm sure you can figure it out (the list comprehension is there because i didn't feel  like adding ['pass'] to each value in the list, feel free to manually do that and remove the comprehension. important part is the .count()).